services/user-service/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-jose
httpx

